Crate vtables

Enums

CustomDataAccessError

Traits

VTableAccessor

VirtualClass

Functions

vtable_custom_data

Method for accessing custom data stores created for each vtable

vtable_custom_data_mut

Method for mutably accessing custom data stores created for each vtable

vtable_mutation_guard

Guard to be used when attempting to mutate a vtable’s contents

vtable_read_guard

Guard to be used when attempting to read a vtable’s contents

vtable_restore_vtable

Method for mutably accessing custom data stores created for each vtable

Attribute Macros

vtable