vtables

Function vtable_custom_data

pub fn vtable_custom_data<V, T: VirtualClass + Deref<Target = V>>(
    vtable: &V,
) -> Result<&T::CustomData, CustomDataAccessError>

Method for accessing custom data stores created for each vtable

The storage is created when the vtable is relocated, along with other identifying information.

Panics

The vtable address is null

The vtable address is not aligned

The vtable has not already been relocated

The object is not an instance of T