Function vtables::vtable_mutation_guard

pub fn vtable_mutation_guard<V, T: VirtualClass + DerefMut<Target = V>>(
    vtable: &mut V
)

Guard to be used when attempting to mutate a vtable’s contents

This will perform checks on the vtable to ensure that it has been reallocated into a unique memory location where mutation is safe.

Compile-time Safety

This function does depend on the implementor of the VirtualClass trait to know how many entries there are in the vtable.

Using the smashline vtable API, you’ll notice that there should be a 1-1 correlation between the types implementing the VirtualClass struct and the offset of the vtable.

This can be violated by the programmer at compile time, however hopefully this documentation should serve as a sufficient troubleshooting resource.

Runtime Safety

The methodology for vtable hijacking done here is designed to be safe at runtime, where panics will occur instead of undefined behavior.

The T type parameter is passed in as a form of type safety:

If this vtable has not yet been mutated, then the address of the vtable is compared against the offset provided in the VirtualClass trait implementation. If successful, the vtable will be relocated and the type id of T will be stored in the additional data store provided by this API

If this vtable has been mutated, then the type if of T will be compared to the one stored at mutation time. If they do not match, then this function will panic.