vtables
Function vtable_read_guard

pub fn vtable_read_guard<V, T: VirtualClass + Deref<Target = V>>(vtable: &V)

Guard to be used when attempting to read a vtable’s contents

This method has the same stipulations and safety advice as vtable_mutation_guard, with the key difference being that this method will not mutate or relocate the vtable in any way.